Arrival.refine_with_trailing_unpack
This new planning refinement method replaces a bunch of
Arrivals by the Arrival of a pack, followed by an Unpack.

At a lower level, Unpack now features plan_for_outcomes() which
creates a planned Unpack with some outcomes already known. It is not
necessary that all of these outcomes are covered by the Unpack nor that
the outcomes of the Unpack be a subset of these outcomes, as the tests
demonstrate.

Unpack's forward_props() is been converted in this commit to
outcome_props_update() which only computes a dict representing an update
to perform, so that it can be used both in normal creation and in
plan_for_outcomes()