Inventory Action compute: restricting Lines to current Node
Forgot previously to restrict the Lines compared to existing
PhysObjs and Avatars to those of the current Node!

Since we only tried compute_actions() with only one None it
didn't show up in tests before.

As for the restriction itself, I also tried inside the join
condition, but that doesn't work, mostly making it a left join
(even with the OR IS NULL).

I just hope the current query doesn't translate to PG as
performing the full OUTER join for all existing lines, and only then
restricting. That would be problematic for performance and even
consistency.